feat: Add CandidateProfileRepository and RecruiterProfileRepository
These code changes add the `CandidateProfileRepository` and `RecruiterProfileRepository` interfaces. These repositories are responsible for handling database operations related to the `CandidateProfile` and `RecruiterProfile` entities respectively. The repositories extend the `JpaRepository` interface and provide methods for CRUD operations on the corresponding entities. This addition enables the application to interact with the database and perform operations on candidate and recruiter profiles.