feat: Add @nextui-org/radio package and update @nextui-org/select version
This commit adds the "@nextui-org/radio" package to the project dependencies, allowing the application to use radio buttons for user input. Additionally, it updates the version of the "@nextui-org/select" package to "^2.2.1" to ensure compatibility with other components. These changes enhance the user interface and improve the selection functionality in the application.